Improve module hierarchy of `custom_shader` example
