+++ examples/custom_shader/src/scene/camera.rs
