+++ examples/custom_shader/src/scene/pipeline.rs
@@ -1,6 +1,15 @@
-use crate::primitive;
-use crate::primitive::cube;
-use crate::primitive::{Buffer, Uniforms};
+pub mod cube;
+
+mod buffer;
+mod uniforms;
+mod vertex;
+
+pub use cube::Cube;
+pub use uniforms::Uniforms;
+
+use buffer::Buffer;
+use vertex::Vertex;
+
 use crate::wgpu;
 use crate::wgpu::util::DeviceExt;
 
@@ -221,7 +230,7 @@ impl Pipeline {
             device.create_shader_module(wgpu::ShaderModuleDescriptor {
                 label: Some("cubes shader"),
                 source: wgpu::ShaderSource::Wgsl(std::borrow::Cow::Borrowed(
-                    include_str!("shaders/cubes.wgsl"),
+                    include_str!("../shaders/cubes.wgsl"),
                 )),
             });
 
@@ -232,7 +241,7 @@ impl Pipeline {
                 vertex: wgpu::VertexState {
                     module: &shader,
                     entry_point: "vs_main",
-                    buffers: &[primitive::Vertex::desc(), cube::Raw::desc()],
+                    buffers: &[Vertex::desc(), cube::Raw::desc()],
                 },
                 primitive: wgpu::PrimitiveState::default(),
                 depth_stencil: Some(wgpu::DepthStencilState {
@@ -468,7 +477,7 @@ impl DepthPipeline {
             device.create_shader_module(wgpu::ShaderModuleDescriptor {
                 label: Some("cubes.depth_pipeline.shader"),
                 source: wgpu::ShaderSource::Wgsl(std::borrow::Cow::Borrowed(
-                    include_str!("shaders/depth.wgsl"),
+                    include_str!("../shaders/depth.wgsl"),
                 )),
             });
 
@@ -573,12 +582,12 @@ impl DepthPipeline {
 }
 
 fn load_skybox_data() -> Vec<u8> {
-    let pos_x: &[u8] = include_bytes!("textures/skybox/pos_x.jpg");
-    let neg_x: &[u8] = include_bytes!("textures/skybox/neg_x.jpg");
-    let pos_y: &[u8] = include_bytes!("textures/skybox/pos_y.jpg");
-    let neg_y: &[u8] = include_bytes!("textures/skybox/neg_y.jpg");
-    let pos_z: &[u8] = include_bytes!("textures/skybox/pos_z.jpg");
-    let neg_z: &[u8] = include_bytes!("textures/skybox/neg_z.jpg");
+    let pos_x: &[u8] = include_bytes!("../textures/skybox/pos_x.jpg");
+    let neg_x: &[u8] = include_bytes!("../textures/skybox/neg_x.jpg");
+    let pos_y: &[u8] = include_bytes!("../textures/skybox/pos_y.jpg");
+    let neg_y: &[u8] = include_bytes!("../textures/skybox/neg_y.jpg");
+    let pos_z: &[u8] = include_bytes!("../textures/skybox/pos_z.jpg");
+    let neg_z: &[u8] = include_bytes!("../textures/skybox/neg_z.jpg");
 
     let data: [&[u8]; 6] = [pos_x, neg_x, pos_y, neg_y, pos_z, neg_z];
 
@@ -597,7 +606,7 @@ fn load_skybox_data() -> Vec<u8> {
 }
 
 fn load_normal_map_data() -> Vec<u8> {
-    let bytes: &[u8] = include_bytes!("textures/ice_cube_normal_map.png");
+    let bytes: &[u8] = include_bytes!("../textures/ice_cube_normal_map.png");
 
     image::load_from_memory_with_format(bytes, image::ImageFormat::Png)
         .unwrap()
