+++ examples/custom_shader/src/scene/pipeline/buffer.rs
