+++ examples/custom_shader/src/scene/pipeline/cube.rs
@@ -1,4 +1,4 @@
-use crate::primitive::Vertex;
+use crate::scene::pipeline::Vertex;
 use crate::wgpu;
 
 use glam::{vec2, vec3, Vec3};
