+++ examples/custom_shader/src/scene/pipeline/vertex.rs
